requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.0
alembic==1.10.3
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
autopage==0.5.1
backcall==0.2.0
black==23.3.0
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
cliff==4.2.0
cmaes==0.9.1
cmake==3.26.1
cmd2==2.4.3
colorlog==6.7.0
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.6
docker-pycreds==0.4.0
docopt==0.6.2
exceptiongroup==1.1.1
executing==1.2.0
filelock==3.11.0
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
ftfy==6.1.1
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
huggingface-hub==0.13.4
hydra-colorlog==1.2.0
hydra-core==1.2.0
hydra-optuna-sweeper==1.2.0
identify==2.5.22
idna==3.4
imageio==2.27.0
importlib-metadata==6.5.0
iniconfig==2.0.0
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.1.0
jupyter_core==5.3.0
kiwisolver==1.4.4
lazy_loader==0.2
lightning-utilities==0.8.0
lit==16.0.0
Mako==1.2.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
monai==1.1.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
nodeenv==1.7.0
num2words==0.5.12
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
open-clip-torch==2.17.1
opencv-python-headless==4.7.0.72
optuna==2.10.1
packaging==23.0
pandas==2.0.0
parso==0.8.3
pathspec==0.11.1
pathtools==0.1.2
pbr==5.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
pluggy==1.0.0
pre-commit==3.2.2
prettytable==3.7.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.0
pyparsing==3.0.9
pyperclip==1.8.2
pyrootutils==1.0.4
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytorch-lightning==2.0.0
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.2
qudida==0.0.4
regex==2023.3.23
requests==2.31.0
rich==13.3.4
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
sentencepiece==0.1.97
sentry-sdk==1.20.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
SQLAlchemy==2.0.9
stack-data==0.6.2
stevedore==5.0.0
sympy==1.11.1
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.6.13
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.0
torchmetrics==0.10.0
torchvision==0.15.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.2
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
virtualenv==20.21.0
wandb==0.14.2
wcwidth==0.2.6
yarl==1.8.2
zipp==3.15.0
